completely restructured the project
- Made each endpoint more modular
