+++ .env.example
@@ -1,6 +1,6 @@
 ACCESS_TOKEN_EXPIRE_MINUTES=30
 ALGORITHM=HS256
-API_V1_STR=/api/v1
+API_PREFIX=/api/v1
 CORS_ORIGINS=["http://localhost:3000","http://localhost:8000"]
 OTLP_ENDPOINT=http://localhost:4317
 PROJECT_NAME="FastAPI Template"
