+++ .gitignore
@@ -273,3 +273,4 @@ public
 .sesskey
 
 docker/volumes/*
+src
