+++ app/api/application.py
@@ -7,8 +7,7 @@ from fastapi.staticfiles import StaticFiles
 
 from app.api import docs, health
 from app.api.lifespan import lifespan_setup
-from app.api.v1.api import api_router as api_router_v1
-from app.api.v2.api import api_router as api_router_v2
+from app.api.router import api_router
 from app.core.config import settings
 from app.core.log import configure_logging
 
@@ -24,7 +23,7 @@ def get_app() -> FastAPI:
         version="0.0.1",
         docs_url=None,
         redoc_url=None,
-        openapi_url=f"{settings.API_V1_STR}/openapi.json",
+        openapi_url=f"{settings.API_PREFIX}/openapi.json",
         lifespan=lifespan_setup,
         default_response_class=JSONResponse,
     )
@@ -39,8 +38,7 @@ def get_app() -> FastAPI:
 
     app.include_router(docs.router)
     app.include_router(health.router)
-    app.include_router(api_router_v1, prefix=settings.API_V1_STR)
-    app.include_router(api_router_v2, prefix=settings.API_V2_STR)
+    app.include_router(api_router, prefix=settings.API_PREFIX)
 
     # Adds static directory. This directory is used to access swagger files.
     app.mount("/static", StaticFiles(directory=APP_ROOT / "static"), name="static")
