+++ app/api/auth/deps.py
@@ -1,0 +1,70 @@
+from typing import Annotated
+
+import jwt
+from fastapi import Depends, HTTPException, Security, status
+from fastapi.security import APIKeyHeader, OAuth2PasswordBearer
+from pydantic import ValidationError
+
+from app.api.auth.schemas import TokenPayload
+from app.api.auth.utils import decode_access_token
+from app.api.users.schemas import User
+from app.api.users.service import UserService
+from app.core.config import settings
+from app.dependencies.repositories import get_user_service
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f"{settings.API_PREFIX}/auth/login")
+api_key_header = APIKeyHeader(name="X-API-Key")
+
+
+def get_current_user(
+    token: Annotated[str, Depends(oauth2_scheme)],
+    user_service: Annotated[UserService, Depends(get_user_service)],
+) -> User:
+    try:
+        payload = decode_access_token(token)
+        token_data = TokenPayload(**payload)
+
+        if token_data.sub is None:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Could not validate credentials",
+                headers={"WWW-Authenticate": "Bearer"},
+            )
+
+    except (jwt.exceptions.PyJWTError, ValidationError) as e:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        ) from e
+
+    user = user_service.get_user(int(token_data.sub))
+
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
+        )
+
+    return user
+
+
+def get_current_user_by_api_key(
+    api_key: Annotated[str, Depends(Security(api_key_header))],
+    user_service: Annotated[UserService, Depends(get_user_service)],
+) -> User:
+    check_api_key = True
+
+    if check_api_key is False:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Missing or invalid API key",
+        )
+
+    user = user_service.get_user_by_api_key(api_key)
+
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
+        )
+
+    return user
