+++ app/api/items/schemas.py
@@ -1,0 +1,28 @@
+from typing import Optional
+
+from pydantic import BaseModel, ConfigDict
+
+
+class ItemBase(BaseModel):
+    name: str
+    description: Optional[str] = None
+    price: float
+
+
+class ItemCreate(ItemBase):
+    pass
+
+
+class ItemUpdate(ItemBase):
+    pass
+
+
+# TODO: These should be in models folder?
+class ItemInDB(ItemBase):
+    id: int
+
+    model_config = ConfigDict(from_attributes=True)
+
+
+class Item(ItemInDB):
+    pass
