--- app/api/v2/endpoints/echo.py
@@ -1,18 +1,0 @@
-from fastapi import APIRouter
-
-from app.models.echo import Message
-
-router = APIRouter()
-
-
-@router.post("/", response_model=Message)
-async def send_echo_message_v2(incoming_message: Message) -> Message:
-    """Sends echo back to user.
-
-    Args:
-        incoming_message: incoming message.
-
-    Returns:
-        message same as the incoming.
-    """
-    return incoming_message
