+++ app/core/config.py
@@ -32,7 +32,7 @@ class LogLevel(StrEnum):
 class Settings(BaseSettings):
     ACCESS_TOKEN_EXPIRE_MINUTES: int = 30
     ALGORITHM: str = "HS256"
-    API_V1_STR: str = "/api/v1"
+    API_PREFIX: str = "/api/v1"
     API_V2_STR: str = "/api/v2"
     CORS_ORIGINS: list[str] = ["*"]
     ENVIRONMENT: str = "dev"
