+++ tests/api/test_redis.py
