--- tests/test_core/__init__.py
@@ -1,1 +1,0 @@
-# Make tests.test_core directory a proper package
